tree is built.
used feedback from the author of tree_iterators_rs here: https://github.com/mr-adult/tree-iterators-rs/issues/8#issuecomment-2888549670

the order of nodes isn't deterministic, but that is understandable as I have not yet accounted for the sibling ordering. Whichever order the children of node 1 are in, their children come through as expected.